+++ packages/openapi_code_builder/lib/openapi_code_builder.dart
@@ -4,7 +4,10 @@ import 'package:openapi_code_builder/src/openapi_code_builder.dart';
 export 'package:openapi_code_builder/src/openapi_code_builder.dart'
     show OpenApiLibraryGenerator, OpenApiCodeBuilderUtils;
 
-Builder openapiCodeBuilder(BuilderOptions options) => OpenApiCodeBuilder(
-      orderDirectives: true,
-      useNullSafetySyntax: true,
-    );
+Builder openapiCodeBuilder(BuilderOptions options) {
+  return OpenApiCodeBuilder(
+    orderDirectives: true,
+    useNullSafetySyntax: true,
+    generateProvider: options.config['generateProvider'] as bool? ?? false,
+  );
+}