+++ packages/openapi_code_builder/lib/src/custom_allocator.dart
@@ -8,6 +8,7 @@ class CustomAllocator implements Allocator {
     'package:openapi_base/openapi_base.dart',
     // https://github.com/google/json_serializable.dart/issues/1115
     'package:json_annotation/json_annotation.dart',
+    'package:freezed_annotation/freezed_annotation.dart',
   ];
 
   final _imports = <String, int>{};